fix CI line-check: extract the /v1/audit replay handlers into their own file
My /v1 additions pushed two files past the 2500-line hard cap:
server_state.c (2512) and server.c (2502).

- server_state.c: moved handle_audit_captures / handle_audit_replay into a
  new server/server_audit_replay_routes.c (they use the global jo_ok/jo_err
  from json_fluent.h and send_and_free from server_state_internal.h, so the
  split is clean). server_state.c -> 2469.
- server.c: the two new dispatch-table entries tipped it to 2502; reclaimed
  two lines by dropping a dead orphan section comment (/* Work queue */,
  which labelled no entries) and the redundant /* Attempt log */ label over
  a self-documenting two-op section. server.c -> 2500.
- Wired the new file into SERVER_SRCS (Makefile) and the CMake SERVER_SRCS.
  aimee-server + the listener-live test link it; server-dispatch keeps its
  own handler stubs (doesn't link it), so no duplicate symbol.

line-check + clang-format-19 clean; full shipping set builds; every
unit-test binary links; D7 + perf gates green.
